Move ToastProvider to hooks folder
