+++ src/hooks/toast/ToastProvider.tsx
@@ -2,8 +2,8 @@
 import React, { useState, useCallback, type ReactNode, useMemo } from "react";
 import ReactDOM from "react-dom";
 import { v4 as uuidv4 } from "uuid";
-import { ToastContext, type Toast } from "../../hooks/toast/toastContext";
-import ToastRenderer from "./ToastRenderer";
+import { ToastContext, type Toast } from "./toastContext";
+import ToastRenderer from "../../components/common/ToastRenderer";
 
 export const ToastProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
   const [toasts, setToasts] = useState<Toast[]>([]);
